[frontend] Add create room dialog as a component
