+++ frontend/app/room/create-room-dialog.tsx
@@ -1,0 +1,55 @@
+"use client";
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+
+export default function CreateRoomDialog() {
+  const onClick = () => {
+    console.log("create room");
+  };
+  return (
+    <Dialog>
+      <TooltipProvider delayDuration={0}>
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <DialogTrigger asChild>
+              <button className="px-2" onClick={onClick}>
+                +
+              </button>
+            </DialogTrigger>
+          </TooltipTrigger>
+          <TooltipContent>Create Room</TooltipContent>
+        </Tooltip>
+      </TooltipProvider>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>Create Room</DialogTitle>
+          <DialogDescription>Let&apos;s create a room</DialogDescription>
+        </DialogHeader>
+        <div>TODO</div>
+        <DialogFooter>
+          <DialogClose asChild>
+            <Button type="button" variant="secondary">
+              Cancel
+            </Button>
+          </DialogClose>
+          <Button>Create</Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+}
