+++ frontend/app/room/rooms-sidebar.tsx
@@ -1,61 +1,8 @@
 "use client";
 import type { RoomEntity } from "@/app/lib/dtos";
 import { Stack } from "@/components/layout/stack";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 import { usePathname, useRouter } from "next/navigation";
-
-function CreateRoomButton() {
-  const onClick = () => {
-    console.log("create room");
-  };
-  return (
-    <Dialog>
-      <TooltipProvider delayDuration={0}>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <DialogTrigger asChild>
-              <button className="px-2" onClick={onClick}>
-                +
-              </button>
-            </DialogTrigger>
-          </TooltipTrigger>
-          <TooltipContent>Create Room</TooltipContent>
-        </Tooltip>
-      </TooltipProvider>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Create Room</DialogTitle>
-          <DialogDescription>Let&apos;s create a room</DialogDescription>
-        </DialogHeader>
-        <div>TODO</div>
-        <DialogFooter>
-          <DialogClose asChild>
-            <Button type="button" variant="secondary">
-              Cancel
-            </Button>
-          </DialogClose>
-          <Button>Create</Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
+import CreateRoomDialog from "./create-room-dialog";
 
 function RoomButton({
   room,
@@ -91,7 +38,7 @@ export default function RoomsSidebar({ rooms }: { rooms: RoomEntity[] }) {
     <div className="overflow-y-auto shrink-0 basis-48 pb-4">
       <div className="flex justify-between">
         <div className={`font-bold`}>Chats</div>
-        <CreateRoomButton />
+        <CreateRoomDialog />
       </div>
       <Stack space="space-y-0">
         {rooms.map((room) => (
